feat: enhance account type handling with AccountTypes and introduce AccountTypeDropdown component
