+++ src/features/Accounts/AccountsView.tsx
@@ -1,6 +1,6 @@
 "use client"
 import { useState, useEffect } from "react";
-import { AccountBank, AccountModalAction, AccountsDisplay } from "@/shared/types/accounts.types";
+import { AccountBank, AccountModalAction, AccountsDisplay, AccountTypes } from "@/shared/types/accounts.types";
 import { Account } from "./Accounts";
 import { getAccountProvider } from "@/shared/lib/accounts.lib";
 import { formatNumberToCurrency } from "@/shared/utils/formatNumberCurrency.utils";
@@ -40,7 +40,8 @@ export const AccountsView = ({ accounts }: AccountsViewProps) => {
         accountId: account._id,
         name: account.title,
         amount: formatNumberToCurrency(account.amount),
-        type: account.accountType,
+        // We're sure the account type is not other string than type AccountTypes
+        type: (account.accountType as AccountTypes),
         accountProvider: getAccountProvider(account.accountProvider) // Default to mastercard if not provided
       }))
       setAccountsDisplay(formattedAccounts)
